% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_to_mu.R
\name{prop_to_mu_two_arm}
\alias{prop_to_mu_two_arm}
\title{Convert a proportion of dropouts to a dropout rate.}
\usage{
prop_to_mu_two_arm(prop_withdrawn = 0.2, dco = 28, recruitment, model)
}
\arguments{
\item{{dco}}{Time of data cut-off.}

\item{{recruitment}}{List of recruitment information. 
Containing \enumerate{
              \item Sample size on control, \code{n_0} 
              \item Sample size on experimental, \code{n_1} 
              \item Recruitment period, \code{r_period}
              \item Recruitment parameter for power model, \code{k} 
            }}

\item{{model}}{The piecewise hazard model.
A list containing the \code{change_points} and \code{lambdas}.}
}
\value{
The dropout rate \code{mu} that would lead to \code{prop_withdrawn} at \code{dco}.
}
\description{
\code{prop_to_mu_two_arm} converts a proportion of patients who have withdrawn at data cut-off \code{dco}
  to the corresponding exponential dropout rate. This takes into acount the competing 
  risk of the event-of-interest.
}
