% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_surv.R
\name{surv_pieces}
\alias{surv_pieces}
\title{Probability of event-free at time t, under piecewise constant hazard.}
\usage{
surv_pieces(t, change_points, lambdas, mu = 0, dropouts = FALSE)
}
\arguments{
\item{{t}}{time}

\item{{change_points}}{The breakpoints for the piecewise hazard function.}

\item{{lambdas}}{A vector of length one more than \code{change_points}. 
The hazard between consecutive breakpoints in \code{c(0, change_points, Inf)}.}

\item{{mu}}{The dropout hazard (a constant). Assumed independent to event hazard.}

\item{{dropouts}}{Is dropout the event of interest? Default is FALSE.}
}
\value{
The probability of being free of the event-of-interest at time \code{t}.
}
\description{
\code{surv_pieces} returns the probability of being event-free at time \code{t} assuming piecewise constant hazard.
}
